requirement.txt
pypdf
pdfplumber
sentence-transformers
faiss-cpu
openai
python-dotenv
streamlit
